Fix WebGPU demo shader errors and add IBL support
- Fix implicit conversion warning in timesamples.hh (uint64_t to size_t)
- Fix WGSL shader error: sample all textures unconditionally to satisfy
  uniform control flow requirement for textureSample
- Add IBL (Image-Based Lighting) shader with PBR support
- Add BRDF LUT generation using compute shader (rgba16float format)
- Add HDR environment map loading with robust parser
- Add fallback environment when HDR loading fails
- Fix environment map paths to use actual assets
- Add float-to-half conversion for HDR textures

+++ src/timesamples.hh
@@ -2824,7 +2824,7 @@ struct TimeSamples {
   }
 
   static size_t get_value_array_index(uint64_t ref) {
-    return ref & ~VALUE_ARRAY_DEDUP_BIT;
+    return static_cast<size_t>(ref & ~VALUE_ARRAY_DEDUP_BIT);
   }
 };
